requirements.txt
numpy==1.25.0
scipy==1.11.2
sounddevice==0.4.3
matplotlib==3.7.2
